src/pyobcomp/comparator.py
"""
Main ObjectComparator class for PyObComp.
"""

from typing import Dict, Any, Optional, Union
import yaml
from .config import ToleranceConfig, FieldConfig, ComparisonOptions, ComparisonResult, Field, ComparisonStatus


class ObjectComparator:
    """Main class for object comparison with tolerance settings."""
    
    def __init__(
        self,
        tolerances: Optional[Dict[str, Union[ToleranceConfig, FieldConfig]]] = None,
        options: Optional[ComparisonOptions] = None
    ):
        """Initialize the comparator.
        
        Args:
            tolerances: Dictionary mapping field paths to tolerance/field configs
            options: Global comparison options
        """
        self.tolerances = tolerances or {}
        self.options = options or ComparisonOptions()
    
    @classmethod
    def from_yaml(cls, config_path: str) -> 'ObjectComparator':
        """Create comparator from YAML configuration file.
        
        Args:
            config_path: Path to YAML configuration file
            
        Returns:
            Configured ObjectComparator instance
        """
        with open(config_path, 'r') as f:
            config = yaml.safe_load(f)
        
        # TODO: Parse YAML config into tolerances and options
        return cls()
    
    def compare(self, expected: Any, actual: Any) -> ComparisonResult:
        """Compare two objects.
        
        Args:
            expected: Expected object
            actual: Actual object
            
        Returns:
            ComparisonResult with detailed comparison information
        """
        # TODO: Implement actual comparison logic
        return ComparisonResult(
            matches=True,
            summary="Comparison not yet implemented",
            fields=[]
        )
