requirements.txt
# Core dependencies
PyYAML>=5.0
